+++ NEWS.md
@@ -1,3 +1,24 @@
+# pedtools 2.10.0
+
+* New function `addSibling()` for adding a sibling to an existing pedigree member.
+
+* New function `isHomozygous()` for finding and counting homozygous genotypes in the marker data.
+
+* New function `sameGenotype()` for finding and counting markers for which two individuals have the same genotype.
+
+* New functions `expectedHomozygosity()` and `expectedHeterozygosity()` for computing these quantities from allele frequencies (or marker objects).
+
+* Extracting sub-pedigrees with `subset()` is now more flexible, allowing disconnected outputs. It also has a new argument `missingParents` controlling how to handle parents that are not included in the subset.
+
+* The `branch()` behaves more intuitively thanks to the improved `subset.ped()`.
+
+* `alleles(x, marker)` now accepts multiple markers at once, returning a list of character vectors.
+
+## Other
+
+* `father(x, id)` and `mother(x, id)` now output `NA` when `id` is a founder in `x`. (Previously this gave `character(0)`.)
+
+
 # pedtools 2.9.0
 
 ## New features
@@ -9,7 +30,6 @@
 * New function `removeGenotypes()`, equivalent to (but more intuitive than) `setAlleles(..., alleles = 0)`.
 
 * `getGenotypes()` gains argument `Xchrom` to be used when it is known in advance that all/none markers are on X.
-
 
 ## Minor fixes